# #451 string comparison related performance improvements