--- a/Jint/Native/Function/FunctionInstance.cs
+++ b/Jint/Native/Function/FunctionInstance.cs
@@ -1,11 +1,18 @@
-﻿using Jint.Native.Object;
+﻿using System.Collections.Generic;
+using Jint.Native.Object;
 using Jint.Runtime;
+using Jint.Runtime.Descriptors;
 using Jint.Runtime.Environments;
 
 namespace Jint.Native.Function
 {
     public abstract class FunctionInstance : ObjectInstance, ICallable
     {
+        private const string PropertyNamePrototype = "prototype";
+        private IPropertyDescriptor _prototype;
+        private const string PropertyNameLength = "length";
+        private IPropertyDescriptor _length;
+
         private readonly Engine _engine;
 
         protected FunctionInstance(Engine engine, string[] parameters, LexicalEnvironment scope, bool strict) : base(engine)
@@ -24,10 +31,10 @@ namespace Jint.Native.Function
         /// <returns></returns>
         public abstract JsValue Call(JsValue thisObject, JsValue[] arguments);
 
-        public LexicalEnvironment Scope { get; private set; }
-        
-        public string[] FormalParameters { get; private set; }
-        public bool Strict { get; private set; }
+        public LexicalEnvironment Scope { get; }
+
+        public string[] FormalParameters { get; }
+        public bool Strict { get; }
 
         public virtual bool HasInstance(JsValue v)
         {
@@ -36,7 +43,7 @@ namespace Jint.Native.Function
             {
                 return false;
             }
-            
+
             var po = Get("prototype");
             if (!po.IsObject())
             {
@@ -44,10 +51,10 @@ namespace Jint.Native.Function
             }
 
             var o = po.AsObject();
-            
+
             if (o == null)
             {
-                throw new JavaScriptException(_engine.TypeError);    
+                throw new JavaScriptException(_engine.TypeError);
             }
 
             while (true)
@@ -58,6 +65,7 @@ namespace Jint.Native.Function
                 {
                     return false;
                 }
+
                 if (vObj == o)
                 {
                     return true;
@@ -65,13 +73,7 @@ namespace Jint.Native.Function
             }
         }
 
-        public override string Class
-        {
-            get
-            {
-                return "Function";
-            }
-        }
+        public override string Class => "Function";
 
         /// <summary>
         /// http://www.ecma-international.org/ecma-262/5.1/#sec-15.3.5.4
@@ -89,6 +91,81 @@ namespace Jint.Native.Function
             }
 
             return v;
+        }
+
+        public override IEnumerable<KeyValuePair<string, IPropertyDescriptor>> GetOwnProperties()
+        {
+            if (_prototype != null)
+            {
+                yield return new KeyValuePair<string, IPropertyDescriptor>(PropertyNamePrototype, _prototype);
+            }
+            if (_length != null)
+            {
+                yield return new KeyValuePair<string, IPropertyDescriptor>(PropertyNameLength, _length);
+            }
+
+            foreach (var entry in base.GetOwnProperties())
+            {
+                yield return entry;
+            }
+        }
+
+        public override IPropertyDescriptor GetOwnProperty(string propertyName)
+        {
+            if (propertyName == PropertyNamePrototype)
+            {
+                return _prototype ?? PropertyDescriptor.Undefined;
+            }
+            if (propertyName == PropertyNameLength)
+            {
+                return _length ?? PropertyDescriptor.Undefined;
+            }
+
+            return base.GetOwnProperty(propertyName);
+        }
+
+        protected internal override void SetOwnProperty(string propertyName, IPropertyDescriptor desc)
+        {
+            if (propertyName == PropertyNamePrototype)
+            {
+                _prototype = desc;
+            }
+            else if (propertyName == PropertyNameLength)
+            {
+                _length = desc;
+            }
+            else
+            {
+                base.SetOwnProperty(propertyName, desc);
+            }
+        }
+
+        public override bool HasOwnProperty(string propertyName)
+        {
+            if (propertyName == PropertyNamePrototype)
+            {
+                return _prototype != null;
+            }
+            if (propertyName == PropertyNameLength)
+            {
+                return _length != null;
+            }
+
+            return base.HasOwnProperty(propertyName);
+        }
+
+        public override void RemoveOwnProperty(string propertyName)
+        {
+            if (propertyName == PropertyNamePrototype)
+            {
+                _prototype = null;
+            }
+            if (propertyName == PropertyNameLength)
+            {
+                _prototype = null;
+            }
+
+            base.RemoveOwnProperty(propertyName);
         }
     }
 }