… Jint/Runtime/Environments/DeclarativeEnvironmentRecord.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/MruPropertyCache2.cs